requirements.txt
django
requests
pillow
djangorestframework
psycopg2-binary